+++ spec/rubocop/cop/lint/safe_navigation_with_empty_spec.rb
@@ -4,10 +4,14 @@ RSpec.describe RuboCop::Cop::Lint::SafeNavigationWithEmpty do
   subject(:cop) { described_class.new }
 
   context 'in a conditional' do
-    it 'registers an offense on `&.empty?`' do
+    it 'registers an offense and corrects on `&.empty?`' do
       expect_offense(<<~RUBY)
         return unless foo&.empty?
                       ^^^^^^^^^^^ Avoid calling `empty?` with the safe navigation operator in conditionals.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        return unless foo && foo.empty?
       RUBY
     end